Adding validation for existing sales during product deletion.

+++ src/products/products.service.ts
@@ -1,14 +1,20 @@
-import { Injectable, NotFoundException } from "@nestjs/common";
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from "@nestjs/common";
 import { ProductsRepository } from "./products.repository";
 import { CreateProductDto } from "./dtos/createProduct.dto";
 import { CategoriesRepository } from "src/categories/categories.repository";
 import { UpdateProductDto } from "./dtos/updateProduct.dto";
+import { SalesRepository } from "src/sales/sales.repository";
 
 @Injectable()
 export class ProductsService {
   constructor(
     private readonly productsRepository: ProductsRepository,
     private readonly categoriesRepository: CategoriesRepository,
+    private readonly salesRepository: SalesRepository,
   ) {}
 
   async createProduct(data: CreateProductDto) {
@@ -63,6 +69,12 @@ export class ProductsService {
     try {
       const product = await this.productsRepository.findProductById(productId);
       if (!product) throw new NotFoundException("Product does not exists.");
+
+      const sales = await this.salesRepository.findSalesByProduct(productId);
+      if (sales?.length > 0)
+        throw new BadRequestException(
+          "Cannot delete product, sales record are found for this product.",
+        );
 
       await this.productsRepository.deleteProduct(productId);
